requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.17
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
urllib3==1.26.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
streamlit==1.28.1
tiktoken==0.5.1
pandas==2.0.3
torch==2.0.1
faiss-cpu==1.7.4
pydantic==2.5.0
typing-extensions==4.8.0
scipy==1.11.4
tenacity==8.2.3 